fix: split out includeMatchAncestors, cleanup
