+++ src/views/product-integration/component-view/components/searchable-variable-group-list/helpers/index.ts
@@ -1,2 +1,3 @@
-export * from './apply-variable-query-filter'
-export * from './apply-variable-required-filter'
+export * from './apply-query-filter'
+export * from './apply-required-filter'
+export * from './include-match-ancestors'
